test-requirements.txt
ddt>=0.8.0,<1.1.1
coverage>=4.0.3,<5.0.0
django_nose>=1.3.0,<=2.0.0
edx-lint>=0.5.2,<1.0.0
factory-boy>=2.6.0,<3.0.0
httpretty>=0.8.14,<1.0.0
mock
nose>=1.3.7,<2.0.0
pep8>=1.7.0,<2.0.0
tox>=2.3.1,<3.0.0

